Refactor the code to use a single function to parse the configuration.
The config parser should be placed in a independent module to avoid the
mutual import issues.
